+++ FLySMPS/diodebridge.h
@@ -5,8 +5,8 @@
 ////////////////////////////////////////////////////////////
 #ifndef DIODEBRIDGE_H
 #define DIODEBRIDGE_H
-//#include <stdint.h>
-#include <math.h>
+#include <cmath>
+#include <cstdint>
 
 class DiodeBridge
 {
@@ -14,18 +14,18 @@ public:
     DiodeBridge();
 
     double IDiodePeak(double ICBulckPeak, double ILoadMax);
-    double IDiodeRMS(double ILoadAVG, int FLine, double ConductTime);
+    double IDiodeRMS(double ILoadAVG, int8_t FLine, double ConductTime);
     double IDiodeAVG(double ILoadAVG);
 
-    double IDiodeRMSTot(double ILoadAVG, int FLine, double ConductTime);
+    double IDiodeRMSTot(double ILoadAVG, int8_t FLine, double ConductTime);
 
-    double ILoadAVG(double IDiodePeak, int FLine, double ConductTime);
+    double ILoadAVG(double IDiodePeak, int8_t FLine, double ConductTime);
     
     double DiodeCurrentSlope(double IDiodePeak, double ILoad, double TotCharTime);
     double DiodeConductTime(double IDiodePeak, double DiodeCurrentSlope);
 
-    double MinPeakInVoltage(int VACmin);
-    double MaxPeakInVoltage(int VACmax);
+    double MinPeakInVoltage(int8_t VACmin);
+    double MaxPeakInVoltage(int8_t VACmax);
 };
 
 #endif // DIODEBRIDGE_H
